Update Register.java
Before iteration 2 the Register class was not up to date. The bugs fixed were 1. once the user clicked "submit" to confirm the registration process a new register page would open, this no longer happens. 2. When the user clicks submit the register page would not close, this too now does not happen.